--- src/modules/core/sagas/utils/ddb/index.js
@@ -1,6 +1,0 @@
-/* @flow */
-
-import getDDB from './getDDB';
-
-// eslint-disable-next-line import/prefer-default-export
-export { getDDB };
