+++ src/modules/core/sagas/utils/getDDB.js
@@ -4,9 +4,9 @@ import { call, getContext } from 'redux-saga/effects';
 
 import { create } from '~utils/saga/effects';
 
-import { DDB as DDBClass } from '../../../../../lib/database/index';
+import { DDB as DDBClass } from '../../../../lib/database/index';
 // eslint-disable-next-line max-len
-import PurserIdentityProvider from '../../../../../lib/database/PurserIdentityProvider';
+import PurserIdentityProvider from '../../../../lib/database/PurserIdentityProvider';
 
 export default function* getDDB(): Generator<*, DDBClass, *> {
   const wallet = yield getContext('wallet');
