--- src/modules/core/sagas/utils/networkClient/index.js
@@ -1,6 +1,0 @@
-/* @flow */
-
-import getNetworkClient from './getNetworkClient';
-
-// eslint-disable-next-line import/prefer-default-export
-export { getNetworkClient };
